+++ frappe/email/__init__.py
@@ -1,7 +1,12 @@
 # Copyright (c) 2015, Frappe Technologies Pvt. Ltd. and Contributors
 # License: MIT. See LICENSE
 
+from typing import TYPE_CHECKING, Literal, Optional
+
 import frappe
+
+if TYPE_CHECKING:
+	from frappe.email.doctype.email_queue.email_queue import EmailQueue
 
 
 def sendmail_to_system_managers(subject, content):
@@ -92,3 +97,137 @@ def get_communication_doctype(doctype, txt, searchfield, start, page_len, filter
 		]
 
 	return [[dt] for dt in com_doctypes if txt.lower().replace("%", "") in dt.lower() and dt in can_read]
+
+
+def sendmail(
+	recipients=None,
+	sender="",
+	subject="No Subject",
+	message="No Message",
+	as_markdown=False,
+	delayed=True,
+	reference_doctype=None,
+	reference_name=None,
+	unsubscribe_method=None,
+	unsubscribe_params=None,
+	unsubscribe_message=None,
+	add_unsubscribe_link=1,
+	attachments=None,
+	content=None,
+	doctype=None,
+	name=None,
+	reply_to=None,
+	queue_separately=False,
+	cc=None,
+	bcc=None,
+	message_id=None,
+	in_reply_to=None,
+	send_after=None,
+	expose_recipients=None,
+	send_priority=1,
+	communication=None,
+	retry=1,
+	now=None,
+	read_receipt=None,
+	is_notification=False,
+	inline_images=None,
+	template=None,
+	args=None,
+	header=None,
+	print_letterhead=False,
+	with_container=False,
+	email_read_tracker_url=None,
+	x_priority: Literal[1, 3, 5] = 3,
+	email_headers=None,
+) -> Optional["EmailQueue"]:
+	"""Send email using user's default **Email Account** or global default **Email Account**.
+
+
+	:param recipients: List of recipients.
+	:param sender: Email sender. Default is current user or default outgoing account.
+	:param subject: Email Subject.
+	:param message: (or `content`) Email Content.
+	:param as_markdown: Convert content markdown to HTML.
+	:param delayed: Send via scheduled email sender **Email Queue**. Don't send immediately. Default is true
+	:param send_priority: Priority for Email Queue, default 1.
+	:param reference_doctype: (or `doctype`) Append as communication to this DocType.
+	:param reference_name: (or `name`) Append as communication to this document name.
+	:param unsubscribe_method: Unsubscribe url with options email, doctype, name. e.g. `/api/method/unsubscribe`
+	:param unsubscribe_params: Unsubscribe paramaters to be loaded on the unsubscribe_method [optional] (dict).
+	:param attachments: List of attachments.
+	:param reply_to: Reply-To Email Address.
+	:param message_id: Used for threading. If a reply is received to this email, Message-Id is sent back as In-Reply-To in received email.
+	:param in_reply_to: Used to send the Message-Id of a received email back as In-Reply-To.
+	:param send_after: Send after the given datetime.
+	:param expose_recipients: Display all recipients in the footer message - "This email was sent to"
+	:param communication: Communication link to be set in Email Queue record
+	:param inline_images: List of inline images as {"filename", "filecontent"}. All src properties will be replaced with random Content-Id
+	:param template: Name of html template from templates/emails folder
+	:param args: Arguments for rendering the template
+	:param header: Append header in email
+	:param with_container: Wraps email inside a styled container
+	:param x_priority: 1 = HIGHEST, 3 = NORMAL, 5 = LOWEST
+	:param email_headers: Additional headers to be added in the email, e.g. {"X-Custom-Header": "value"} or {"Custom-Header": "value"}. Automatically prepends "X-" to the header name if not present.
+	"""
+
+	from frappe.utils.jinja import get_email_from_template
+
+	if recipients is None:
+		recipients = []
+	if cc is None:
+		cc = []
+	if bcc is None:
+		bcc = []
+
+	text_content = None
+	if template:
+		message, text_content = get_email_from_template(template, args)
+
+	message = content or message
+
+	if as_markdown:
+		from frappe.utils import md_to_html
+
+		message = md_to_html(message)
+
+	if not delayed:
+		now = True
+
+	from frappe.email.doctype.email_queue.email_queue import QueueBuilder
+
+	builder = QueueBuilder(
+		recipients=recipients,
+		sender=sender,
+		subject=subject,
+		message=message,
+		text_content=text_content,
+		reference_doctype=doctype or reference_doctype,
+		reference_name=name or reference_name,
+		add_unsubscribe_link=add_unsubscribe_link,
+		unsubscribe_method=unsubscribe_method,
+		unsubscribe_params=unsubscribe_params,
+		unsubscribe_message=unsubscribe_message,
+		attachments=attachments,
+		reply_to=reply_to,
+		cc=cc,
+		bcc=bcc,
+		message_id=message_id,
+		in_reply_to=in_reply_to,
+		send_after=send_after,
+		expose_recipients=expose_recipients,
+		send_priority=send_priority,
+		queue_separately=queue_separately,
+		communication=communication,
+		read_receipt=read_receipt,
+		is_notification=is_notification,
+		inline_images=inline_images,
+		header=header,
+		print_letterhead=print_letterhead,
+		with_container=with_container,
+		email_read_tracker_url=email_read_tracker_url,
+		x_priority=x_priority,
+		email_headers=email_headers,
+	)
+
+	# build email queue and send the email if send_now is True.
+	return builder.process(send_now=now)
